✨Cicle counter working,still need to be testedmain

+++ libraries/ciclo/Encoder.hpp
@@ -21,7 +21,7 @@ class Encoder {
         int A_pin;
         int B_pin;
         int PulsesPerRev;
-        int PitchPerRev;
+        float PitchPerRev;
         int pulses=0;
         int mode;
         float k;
@@ -49,7 +49,7 @@ class Encoder {
          * @param A_pin A phase pin 
          * @param B_pin B phase pin 
          */
-        Encoder(int A_pin,int B_pin, const byte which,int PulsesPerRev,int PitchPerRev,int mode);
+        Encoder(int A_pin,int B_pin, const byte which,int PulsesPerRev,float PitchPerRev,int mode);
 
         void init();
